bump to v38.0.0

+++ include/autd3/native_methods/autd3-core.h
@@ -112,7 +112,7 @@ struct PulseWidthInner {
   };
 
   struct Raw_Body {
-    uint32_t _0;
+    uint16_t _0;
   };
 
   Tag tag;
